% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{build_weight_lookup}
\alias{build_weight_lookup}
\title{Pre-build Edge-Weight Lookup Matrices}
\usage{
build_weight_lookup(instance, n, num_obj)
}
\arguments{
\item{instance}{\code{data.frame} returned by \code{\link{generate_instance}}
or one with columns \code{from}, \code{to}, \code{weight_1}, \code{weight_2}
(and optionally \code{weight_3}).}

\item{n}{Integer. Number of nodes.}

\item{num_obj}{Integer. Number of objectives (2 or 3).}
}
\value{
A list of length \code{num_obj}; element \code{k} is the symmetric
  \code{n x n} weight matrix of objective \code{k}.
}
\description{
Constructs one \code{n x n} matrix per objective so that the weight of any
edge \code{(i, j)} can be queried in O(1) via standard matrix indexing.
}
\examples{
inst <- generate_instance(10, 2, seed = 1)
L <- build_weight_lookup(inst, 10, 2)
L[[1]][1, 2]
}
