% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prufer.R
\name{generate_prufer_population}
\alias{generate_prufer_population}
\title{Generate an Initial Prufer-Encoded Population}
\usage{
generate_prufer_population(n, pop_size)
}
\arguments{
\item{n}{Integer. Number of nodes.}

\item{pop_size}{Integer. Number of individuals to generate.}
}
\value{
Integer matrix of dimensions \code{pop_size x (n - 2)}.
}
\description{
Generate an Initial Prufer-Encoded Population
}
\examples{
generate_prufer_population(n = 6, pop_size = 4)
}
