% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsga2_helpers.R
\name{non_dominated_crowding}
\alias{non_dominated_crowding}
\title{Assign Pareto Rank and Crowding Distance}
\usage{
non_dominated_crowding(population, num_obj)
}
\arguments{
\item{population}{Numeric matrix \code{[N x (vars + num_obj)]}.}

\item{num_obj}{Integer.}
}
\value{
Matrix with extra columns \code{rankingIndex} and \code{density}.
}
\description{
Assign Pareto Rank and Crowding Distance
}
