% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_search.R
\name{pareto_local_search}
\alias{pareto_local_search}
\title{Pareto Local Search}
\usage{
pareto_local_search(
  instance,
  pareto_pop,
  num_obj,
  n,
  neighbour_frac = 0.1,
  pop_size,
  lookup = NULL,
  verbose = FALSE
)
}
\arguments{
\item{instance}{Edge-list \code{data.frame}.}

\item{pareto_pop}{Integer matrix \code{[k x (n - 2)]}.}

\item{num_obj}{Integer.}

\item{n}{Integer.}

\item{neighbour_frac}{Numeric.}

\item{pop_size}{Integer.}

\item{lookup}{Optional lookup.}

\item{verbose}{Logical.}
}
\value{
data.frame of chromosomes.
}
\description{
Pareto Local Search
}
