% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_search.R
\name{path_relinking}
\alias{path_relinking}
\title{Path Relinking on the Current Pareto Front}
\usage{
path_relinking(
  instance,
  pareto_pop,
  num_obj,
  n,
  pop_size,
  lookup = NULL,
  verbose = FALSE
)
}
\arguments{
\item{instance}{Edge-list \code{data.frame}.}

\item{pareto_pop}{Integer matrix \code{[k x (n - 2)]}.}

\item{num_obj}{Integer.}

\item{n}{Integer.}

\item{pop_size}{Integer.}

\item{lookup}{Optional lookup.}

\item{verbose}{Logical.}
}
\value{
data.frame of non-dominated chromosomes.
}
\description{
Path Relinking on the Current Pareto Front
}
