% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_best_tree}
\alias{plot_best_tree}
\title{Plot the Best-Compromise Spanning Tree}
\usage{
plot_best_tree(result, n)
}
\arguments{
\item{result}{List returned by \code{\link{run_momst}}.}

\item{n}{Integer.}
}
\value{
Invisible NULL.
}
\description{
Plot the Best-Compromise Spanning Tree
}
