% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pareto_front}
\alias{plot_pareto_front}
\title{Plot a Pareto Front (2-objective case)}
\usage{
plot_pareto_front(result, show_dominated = FALSE)
}
\arguments{
\item{result}{List returned by \code{\link{run_momst}}.}

\item{show_dominated}{Logical.}
}
\value{
Invisible NULL.
}
\description{
Plot a Pareto Front (2-objective case)
}
