% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{random_mutation}
\alias{random_mutation}
\title{Random Mutation on Prufer Sequences}
\usage{
random_mutation(population, pop_size, mut_rate)
}
\arguments{
\item{population}{Numeric matrix \code{[pop_size x (n - 2)]}.}

\item{pop_size}{Integer.}

\item{mut_rate}{Numeric in \[0, 1\].}
}
\value{
Integer matrix \code{[pop_size x (n - 2)]}.
}
\description{
Random Mutation on Prufer Sequences
}
