% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsga2_helpers.R
\name{tournament_selection}
\alias{tournament_selection}
\title{Tournament Selection}
\usage{
tournament_selection(population, pop_size, tour_size)
}
\arguments{
\item{population}{Matrix with last two columns being \code{rankingIndex}
and \code{density}.}

\item{pop_size}{Integer.}

\item{tour_size}{Integer.}
}
\value{
Selected subpopulation matrix.
}
\description{
Tournament Selection
}
