% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{uniform_crossover}
\alias{uniform_crossover}
\title{Uniform Crossover for Prufer Sequences}
\usage{
uniform_crossover(pool, pop_size, cross_rate)
}
\arguments{
\item{pool}{Numeric matrix \code{[pop_size x (n - 2)]}.}

\item{pop_size}{Integer (must be even).}

\item{cross_rate}{Numeric in \[0, 1\].}
}
\value{
Integer matrix \code{[pop_size x (n - 2)]}.
}
\description{
Uniform Crossover for Prufer Sequences
}
